Simplify package description.
Go style suggests, as verified by `golint`, that Go packages that are libraries (i.e., package name is not "main") package description begin with "Package {{.Name}} ...". I used to think this applied to commands as well, but I've learned that's not the case. There's no explicit requirement to start package descriptions for commands (i.e., Go package with name "main") with "Command {{.Name}} ...".

That allows to simplify the package description by omitting that word.